+++ @rocketseat/gatsby-theme-docs-core/gatsby-config.js
@@ -1,12 +1,20 @@
 const withDefault = require(`./util/with-default`);
 const path = require(`path`);
+const gfmPlugin = require(`remark-gfm`);
+const { rehypeMetaAsAttributes } = require(`./rehype-meta-as-attributes`);
 const camelCase = require('lodash.camelcase');
 
 const upperFirst = (text) => text.charAt(0).toUpperCase() + text.slice(1);
 
 module.exports = (options) => {
-  const { basePath, configPath, docsPath, yamlFilesPath, withMdx } =
-    withDefault(options);
+  const {
+    basePath,
+    configPath,
+    docsPath,
+    yamlFilesPath,
+    withMdx,
+    gatsbyRemarkPlugins,
+  } = withDefault(options);
 
   return {
     siteMetadata: {
@@ -64,6 +72,10 @@ module.exports = (options) => {
         resolve: `gatsby-plugin-mdx`,
         options: {
           extensions: [`.mdx`, `.md`],
+          mdxOptions: {
+            remarkPlugins: [gfmPlugin],
+            rehypePlugins: [rehypeMetaAsAttributes],
+          },
           gatsbyRemarkPlugins: [
             `gatsby-remark-autolink-headers`,
             `gatsby-remark-embedder`,
@@ -77,8 +89,8 @@ module.exports = (options) => {
             },
             `gatsby-remark-responsive-iframe`,
             `gatsby-remark-copy-linked-files`,
+            ...gatsbyRemarkPlugins,
           ],
-          plugins: [`gatsby-remark-autolink-headers`, `gatsby-remark-images`],
         },
       },
     ].filter(Boolean),