+++ @rocketseat/gatsby-theme-docs/gatsby-config.js
@@ -1,7 +1,7 @@
 const withDefault = require('@rocketseat/gatsby-theme-docs-core/util/with-default');
 
 module.exports = (options) => {
-  const themeOtions = withDefault(options);
+  const themeOptions = withDefault(options);
 
   return {
     siteMetadata: {
@@ -11,15 +11,22 @@ module.exports = (options) => {
       siteDescription: `Easy to use Gatsby Theme to create documentation websites`,
       siteUrl: `https://rocketdocs.netlify.app`,
       siteAuthor: `@jpedroschmitz`,
-      siteImage: `/banner.jpg`,
+      siteImage: `/banner.png`,
       siteLanguage: `en`,
-      basePath: themeOtions.basePath,
-      docsPath: themeOtions.docsPath,
+      basePath: themeOptions.basePath,
+      docsPath: themeOptions.docsPath,
     },
     plugins: [
       {
+        resolve: `gatsby-source-filesystem`,
+        options: {
+          name: `homepage`,
+          path: themeOptions.homePath || `${__dirname}/src/text`,
+        },
+      },
+      {
         resolve: `@rocketseat/gatsby-theme-docs-core`,
-        options: themeOtions,
+        options: themeOptions,
       },
       `gatsby-plugin-catch-links`,
       `gatsby-plugin-emotion`,