+++ @rocketseat/gatsby-theme-docs-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rocketseat/gatsby-theme-docs-core
 
+## 4.0.0
+
+### Major Changes
+
+- [#150](https://github.com/jpedroschmitz/rocketdocs/pull/150) [`ed4b5b3`](https://github.com/jpedroschmitz/rocketdocs/commit/ed4b5b31dd60f732abced13d09df97671e9af828) Thanks [@tekmaven](https://github.com/tekmaven)! - Upgrade to Gatsby v5
+
+## 3.2.0
+
+### Minor Changes
+
+- [#146](https://github.com/jpedroschmitz/rocketdocs/pull/146) [`a4903ae`](https://github.com/jpedroschmitz/rocketdocs/commit/a4903ae1f5bb24cdb075b15f374135e7da554511) Thanks [@aaronamm](https://github.com/aaronamm)! - Add `gatsbyRemarkPlugins` option
+
 ## 3.1.0
 
 ### Minor Changes